2024 09: code cleaning

+++ src/year2024/day06.rs
@@ -83,14 +83,14 @@ pub fn solve(input: &str) -> Result<(usize, usize)> {
     let p1 = vseen.len();
 
     let slide = Slide::new(&grid);
-    /* 
+
     let p2 = vseen
         .into_par_iter()
         .map(|(obsx, obsy)| {
             has_cycle(&grid, &slide, (start.0 as i32, start.1 as i32), *obsx as i32, *obsy as i32) as u32 
         })
-        .sum();
-    */
+        .sum() as usize;
+    /* 
     let p2 = vseen
         .into_par_iter()
         .chunks(32)
@@ -103,6 +103,7 @@ pub fn solve(input: &str) -> Result<(usize, usize)> {
 
         })
         .sum();
+    */
 
     Ok((p1, p2))
 }
